Protection Implementations and Bug Fixes
Protected embed() behind a role called "admin"; protected the Discord on_raw_reaction_add method with an if statement; fixed some bugs that were preventing the embed() command from working properly

+++ bot.py
@@ -20,22 +20,25 @@ class TestCog(commands.Cog):
     async def on_raw_reaction_add(self, payload):  # called when a user reacts
         if payload.user_id == bot.user.id:  # Prevent the chat log from being sent to the admin channel before deletion
             return
-
         #TODO: check to see the message is from the bot and it is actually an embed message
         channel = bot.get_channel(payload.channel_id)  # get channel id from payload
         message = await channel.fetch_message(payload.message_id)  # get message id from payload
-        embed = message.embeds[0]  # get the embed from the message
-        footer = b64.decode(embed.footer.text)
 
-        # searches footer["type"] for channel type (help / sponsor specific / delete)
-        if footer["type"] == "HELP_DESK": 
-            await reaction_add.create_help_channel(self, payload, bot)
-        elif footer["type"] == "DELETE_HELP_CHANNEL":
-            await reaction_add.delete_help_channel(self, payload, bot)
+        if len(message.embeds) != 0 and message.author.id == bot.user.id:
+            embed = message.embeds[0]  # get the embed from the message
+
+            footer = b64.decode(embed.footer.text)
+
+            # searches footer["type"] for channel type (help / sponsor specific / delete)
+            if footer["type"] == "HELP_DESK":
+                await reaction_add.create_help_channel(self, payload, bot)
+            elif footer["type"] == "DELETE_HELP_CHANNEL":
+                await reaction_add.delete_help_channel(self, payload, bot)
 
     #TODO: protect this command with a role so that a random user cannot invoke it
     #TODO: change channel_category to be an ID to an existing category, and update it in create_help_channel when searching
     @commands.command()
+    @commands.has_role('admin')
     async def embed(self, ctx, channel_category, channel_name, custom_ticket, user_reaction, *, text):
         await ctx.message.delete()  # immediately deletes original command from chat
         # for customized title, create argument for title, and pass argument into title= 
@@ -50,7 +53,7 @@ class TestCog(commands.Cog):
         footer["ticket_num"] = 0
         footer["type"] = "HELP_DESK"
         
-        # embed.set_footer(text=b64.encode(footer))  # add category to embed footer
+        embed.set_footer(text=b64.encode(footer))  # add category to embed footer
         msg = await ctx.send(file=file, embed=embed)
 
         await msg.add_reaction(user_reaction)
@@ -60,7 +63,7 @@ class TestCog(commands.Cog):
         for category in ctx.message.guild.categories:
             channel_category.replace("_", " ")
             if channel_category == category:
-                found = True 
+                found = True
                 break
 
         # creates category on embed message send
